motor/requirements.txt
aioamqp==0.13.0
aioredis==1.3.1
alabaster==0.7.12
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgi-rabbitmq==0.5.5
asgi-redis==1.4.3
asgiref==3.2.0
astroid==2.3.3
async-timeout==2.0.0
attrs==19.3.0
autobahn==17.6.2
Automat==0.6.0
Babel==2.5.3
bandit==1.6.2
billiard==3.3.0.23
black==19.10b0
bleach==3.1.0
cached-property==1.3.0
celery==3.1.25
certifi==2019.11.28
cffi==1.13.2
channels==2.3.1
channels-rabbitmq==1.1.5
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
coverage==4.5.4
cryptography==2.8
daphne==2.4.0
defusedxml==0.6.0
Django==3.0
django-celery==3.2.1
django-easy-pjax==1.2.0
django-material==0.4.1
django-secure==1.0.1
djangorestframework==3.11.0
docutils==0.15.2
easygui==0.98.1
ecdsa==0.13
entrypoints==0.3
et-xmlfile==1.0.1
fernet==1.0.1
flake8==3.7.9
flup==1.0.3
fpdf2==2.0.3
future==0.16.0
futures==3.1.1
gitdb2==2.0.3
GitPython==2.1.1
gunicorn==19.6.0
hiredis==1.0.1
hyperlink==17.2.1
idna==2.7
imagesize==1.1.0
importlib-metadata==1.3.0
incremental==17.5.0
isort==4.3.21
jdcal==1.4.1
jeepney==0.4.1
Jinja2==2.10.3
json-rpc==1.12.2
jsonfield==1.0.3
keyring==20.0.0
kombu==3.0.37
lazy-object-proxy==1.4.3
lxml==4.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
msgpack==0.6.2
msgpack-python==0.4.8
mypy==0.750
mypy-extensions==0.4.3
ninja==1.9.0.post1
numpy==1.17.4
odfpy==1.4.0
openpyxl==2.6.0
packaging==16.8
pamqp==2.3.0
pathspec==0.6.0
pbr==5.4.4
pg8000==1.13.2
pika==0.11.2
Pillow==6.2.1
pineboo==0.65.9
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.10
psycopg2==2.8.3
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.5.1
pydocstyle==5.0.1
PyExecJS==1.1.0
pyflakes==2.1.1
PyGithub==1.31
Pygments==2.5.2
PyHamcrest==1.9.0
pylint==2.4.4
pylint-json2html==0.1.0
pyOpenSSL==19.1.0
pyparsing==2.2.0
PyQt5==5.13.0
PyQt5-sip==4.19.19
PyQt5-stubs==5.13.1.3
PyReact==0.5.2
pyroma==2.6
pytest==5.3.1
pytest-cov==2.8.1
python-barcode==0.10.0
python-dateutil==2.8.1
python-jose==2.0.2
pytz==2017.2
PyYAML==5.2
readme-renderer==24.0
redis==2.10.6
regex==2015.3.18
reportlab==3.5.32
requests==2.20.0
requests-toolbelt==0.9.1
scramp==1.1.0
SecretStorage==3.1.1
selenium==3.141.0
service-identity==18.1.0
six==1.10.0
smmap2==2.0.3
snowballstemmer==2.0.0
SQLAlchemy==1.3.9
sqlparse==0.3.0
stevedore==1.31.0
testresources==2.0.1
toml==0.10.0
tornado==5.0
tqdm==4.40.2
twine==3.1.1
Twisted==19.10.0
txaio==2.8.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.24.3
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
xmljson==0.2.0
xmltodict==0.9.2
zipp==0.6.0
zope.interface==4.4.2
